move is_deallocate_trivial
