+++ include/boost/json/monotonic_resource.hpp
@@ -12,7 +12,7 @@
 #define BOOST_JSON_MONOTONIC_RESOURCE_HPP
 
 #include <boost/json/detail/config.hpp>
-#include <boost/json/storage_ptr.hpp>
+#include <boost/json/memory_resource.hpp>
 #include <boost/json/detail/monotonic_resource.hpp>
 
 BOOST_JSON_NS_BEGIN
