+++ include/boost/json/storage_ptr.hpp
@@ -21,62 +21,6 @@
 
 BOOST_JSON_NS_BEGIN
 
-/** Return true if a memory resource's deallocate function has no effect.
-
-    This metafunction may be specialized to indicate to
-    the library that calls to the `deallocate` function of
-    a @ref memory_resource have no effect. The implementation
-    will elide such calls when it is safe to do so. By default,
-    the implementation assumes that all memory resources
-    require a call to `deallocate` for each memory region
-    obtained by calling `allocate`.
-
-    @par Example
-
-    This example specializes the metafuction for `my_resource`,
-    to indicate that calls to deallocate have no effect:
-
-    @code
-
-    // Forward-declaration for a user-defined memory resource
-    struct my_resource;
-
-    // It is necessary to specialize the template from
-    // inside the namespace in which it is declared:
-
-    namespace boost {
-    namespace json {
-
-    template<>
-    struct is_deallocate_trivial< my_resource >
-    {
-        static constexpr bool value = true;
-    };
-
-    } // namespace json
-    } // namespace boost
-
-    @endcode
-
-    It is usually not necessary for users to check this trait.
-    Instead, they can call @ref storage_ptr::is_deallocate_trivial
-    to determine if the pointed-to memory resource has a trivial
-    deallocate function.
-
-    @see
-        @ref memory_resource,
-        @ref storage_ptr
-*/
-template<class T>
-struct is_deallocate_trivial
-{
-    /** A bool equal to true if calls to `T::do_deallocate` have no effect.
-
-        The primary template sets `value` to false.
-    */
-    static constexpr bool value = false;
-};
-
 /** A smart pointer to a @ref memory_resource
 
     This container is used to hold a pointer to a
